Created WrappedStreamingBodyPartition()
