+++ pywren/pywren_ibm_cloud/wrenutil.py
@@ -19,6 +19,7 @@ import os
 import uuid
 import inspect
 import subprocess
+import struct
 
 
 def uuid_str():
@@ -149,6 +150,81 @@ def get_current_memory_usage():
     memorybytes = int(process.stdout.read())
 
     return sizeof_fmt(memorybytes)
+
+
+class WrappedStreamingBody:
+    """
+    Wrap boto3's StreamingBody object to provide enough Python fileobj functionality,
+    and to discard data added by partitioner and cut lines.
+
+    from https://gist.github.com/debedb/2e5cbeb54e43f031eaf0
+
+    """
+    def __init__(self, sb, size):
+        # The StreamingBody we're wrapping
+        self.sb = sb
+        # Initial position
+        self.pos = 0
+        # Size of the object
+        self.size = size
+
+    def tell(self):
+        # print("In tell()")
+        return self.pos
+
+    def read(self, n=None):
+        retval = self.sb.read(n)
+        if retval == "":
+            raise EOFError()
+        self.pos += len(retval)
+        return retval
+
+    def readline(self):
+        try:
+            retval = self.sb.readline()
+        except struct.error:
+            raise EOFError()
+        self.pos += len(retval)
+        return retval
+
+    def seek(self, offset, whence=0):
+        # print("Calling seek()")
+        retval = self.pos
+        if whence == 2:
+            if offset == 0:
+                retval = self.size
+            else:
+                raise Exception("Unsupported")
+        else:
+            if whence == 1:
+                offset = self.pos + offset
+                if offset > self.size:
+                    retval = self.size
+                else:
+                    retval = offset
+        # print("In seek(%s, %s): %s, size is %s" % (offset, whence, retval, self.size))
+
+        self.pos = retval
+        return retval
+
+    def __str__(self):
+        return "WrappedBody"
+
+    def __getattr__(self, attr):
+        # print("Calling %s"  % attr)
+
+        if attr == 'tell':
+            return self.tell
+        elif attr == 'seek':
+            return self.seek
+        elif attr == 'read':
+            return self.read
+        elif attr == 'readline':
+            return self.readline
+        elif attr == '__str__':
+            return self.__str__
+        else:
+            return getattr(self.sb, attr)
 
 
 def verify_args(func, data, object_processing=False):
